[Add] Footsteps (both plr and npc), landing, and gun update
Footsteps for the player have a parameter hooked up to the pitch, anything above 11m/s of speed for the player will increase it. You just gotta stick something in the update function to set the player_speed parameter equal to the player's speed and boom, will update when it plays.